fix: not auto-instal example plugins
